+++ Functions/MultiVariableVectorValuedFunction.cpp
@@ -1,72 +1,72 @@
 #include "Functions/MultiVariableVectorValuedFunction.h"
 
 #include "Functions/IRealFunctionUtils.h"
-#include "HLUtils/MessageHandler.h"
+#include "Utilities/MessageHandler.h"
 #include "MatrixVector/MatrixSquare.h"
 
 
 
 MultiVariableVectorValuedFunction::MultiVariableVectorValuedFunction(std::span< std::shared_ptr<IMultiVariableRealValuedFunction> >  components) :
-	m_components(components.begin(), components.end())
+   m_components(components.begin(), components.end())
 {
-	const int numVar = m_components[0]->GetDomainDimension();
-	for (const auto cmp : m_components)
-	{
-		MessageHandler::Assert(cmp->GetDomainDimension() == numVar);
-	}
+   const int numVar = m_components[0]->GetDomainDimension();
+   for (const auto cmp : m_components)
+   {
+      MessageHandler::Assert(cmp->GetDomainDimension() == numVar);
+   }
 }
 
 bool MultiVariableVectorValuedFunction::DerivativeAlwaysZero(int eqn, int var) const
 {
-	const auto cmp = m_components[eqn];
-	return cmp->DerivativeAlwaysZero(0,var);
+   const auto cmp = m_components[eqn];
+   return cmp->DerivativeAlwaysZero(0, var);
 }
 
 
 bool MultiVariableVectorValuedFunction::HasDerivative() const
 {
-	return true;
+   return true;
 }
 
 int MultiVariableVectorValuedFunction::GetRangeDimension() const
 {
-	return m_components.size();
+   return static_cast<int>(m_components.size());
 }
 
 int MultiVariableVectorValuedFunction::GetDomainDimension() const
 {
-	return m_components[0]->GetDomainDimension();
+   return m_components[0]->GetDomainDimension();
 }
 
 void MultiVariableVectorValuedFunction::Evaluate(std::span<const double>x, std::span<double> fx)const
 {
-	int eqn = 0;
-	double eval;
-	std::span<double> evalWrapper(&eval,1);
-	for (const auto cmp : m_components)
-	{
-		cmp->Evaluate(x, evalWrapper);
-		fx[eqn] = eval;
-		++eqn;
-	}
+   int eqn = 0;
+   double eval;
+   std::span<double> evalWrapper(&eval, 1);
+   for (const auto cmp : m_components)
+   {
+      cmp->Evaluate(x, evalWrapper);
+      fx[eqn] = eval;
+      ++eqn;
+   }
 }
 
 
 void MultiVariableVectorValuedFunction::Derivative(std::span<const double>x, IMatrix& dfdx)const
 {
-	const int numVar = GetDomainDimension();
-	const int numEqn = GetRangeDimension();
-	MessageHandler::Assert( dfdx.GetRowDimension() == numEqn);
-	MessageHandler::Assert( dfdx.GetColDimension() == numVar);
-	std::vector<double> deriv(numVar);
-	int eqn = 0;
-	for (const auto cmp : m_components)
-	{
-		 ::Derivative( *cmp, x, deriv);
-		 for (int var = 0; var < numVar; ++var)
-		 {
-			 dfdx(eqn, var) = deriv[var];
-		 }
-		++eqn;
-	}
+   const int numVar = GetDomainDimension();
+   const int numEqn = GetRangeDimension();
+   MessageHandler::Assert(dfdx.GetRowDimension() == numEqn);
+   MessageHandler::Assert(dfdx.GetColDimension() == numVar);
+   std::vector<double> deriv(numVar);
+   int eqn = 0;
+   for (const auto cmp : m_components)
+   {
+      ::Derivative(*cmp, x, deriv);
+      for (int var = 0; var < numVar; ++var)
+      {
+         dfdx(eqn, var) = deriv[var];
+      }
+      ++eqn;
+   }
 }
